Made plugin more robust against different project types.
When the plugin is executed in a project of type pom it throws an
exception because it cannot find the source directory. This is now
suppressed, and no data is generated for this project.

The use case is that there can be multi-module projects defining
multiple tag libs. Then it makes sense to have a parent project (of
type pom) which configures the plugin. The plugin is then applied to
all taglib sub projects. With this patch, this use case is now
supported.

+++ jellydoc-maven-plugin/src/main/java/org/jvnet/maven/jellydoc/JellydocMojo.java
@@ -124,12 +124,23 @@ public class JellydocMojo extends AbstractMojo implements MavenMultiPageReport {
         javadoc.setTaskName("jellydoc");
         javadoc.setProject(p);
 
+        boolean foundDir = false;
         for (Object dir : project.getCompileSourceRoots()) {
-            FileSet fs = new FileSet();
-            fs.setProject(p);
-            fs.setDir(new File(dir.toString()));
-            javadoc.addFileset(fs);
+            File dirFile = new File(dir.toString());
+            // For pom projects the dir might not exist
+            if (dirFile.exists()) {
+                FileSet fs = new FileSet();
+                fs.setProject(p);
+                fs.setDir(dirFile);
+                javadoc.addFileset(fs);
+                foundDir = true;
+            }
         }
+
+        if (!foundDir) {
+            return;
+        }
+
         javadoc.setClasspath(makePath(p,project.getArtifacts()));
 
         Javadoc.DocletInfo d = javadoc.createDoclet();
@@ -245,8 +256,11 @@ public class JellydocMojo extends AbstractMojo implements MavenMultiPageReport {
     public void generate(Sink sink, SinkFactory sinkFactory, Locale locale) throws MavenReportException {
         try {
             execute();
-            new ReferenceRenderer(sink,new File(targetDir(),"taglib.xml").toURI().toURL()).render();
-            FileUtils.copyDirectory(targetDir(),new File(targetDir(),"site"),"taglib-*.xsd",null);
+            File libFile = new File(targetDir(), "taglib.xml");
+            if (libFile.exists()) {
+                new ReferenceRenderer(sink, libFile.toURI().toURL()).render();
+                FileUtils.copyDirectory(targetDir(), new File(targetDir(), "site"), "taglib-*.xsd", null);
+            }
         } catch (AbstractMojoExecutionException | DocumentException | IOException e) {
             throw new MavenReportException("Failed to generate report",e);
         }
